Commands: Warn the user before overwriting tournaments
In the event that a scheduled tournament already exists in the room, the bot will map the user's id to the room's id in the `overwriteWarnings` variable. After that, the bot will request that the user reuse the command to overwrite the existing tournament. This has been tested and works; if you wish to see screenshots, I can provide one.